+++ src/main/java/net/oneandone/pommes/cli/DatabaseAdd.java
@@ -52,7 +52,7 @@ public class DatabaseAdd extends Base {
         this.fixscm = fixscm;
         this.zone = zone;
         this.repositories = new ArrayList<>();
-        this.log = new PrintWriter(environment.logs().join("pommes.log").newWriter(), true);
+        this.log = new PrintWriter(environment.home.logs().join("pommes.log").newWriter(), true);
     }
 
     public void add(String str) throws URISyntaxException, NodeInstantiationException {