+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 * configuration changes
   * POMMES_PROPERTIES replaced by POMMES_HOME
   * dumped `mount.root` property, $POMMES_HOME is used instead
-  * properties file always located in $POMMES_HOME/pommes.properties"
-  * database  always located in $POMMES_HOME/database, removed the `database` property
+  * properties file always located in $POMMES_HOME/.pommes/pommes.properties"
+  * log files go to $POMMES_HOME/.pommes/logs
+  * database always located in $POMMES_HOME/.pommes/database, removed the `database` property
 * dependency updates
 * `ls` improvements
   * renamed to `st`